requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
filelock==3.17.0
fonttools==4.56.0
frozendict==2.4.6
identify==2.6.8
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multitasking==0.0.11
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
virtualenv==20.29.3
yfinance==0.2.55
